Applied [ 880011 ] Several fixes to DbBrowse demo ABX
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@25611 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ demos/dbbrowse/tabpgwin.h
@@ -83,7 +83,7 @@ public:
  wxPen  mDarkPen;         // default: RGB(128,128,128)
  wxPen  mBlackPen;        // default: RGB(  0,  0,  0)
 
- int    mActiveTab;
+ size_t mActiveTab;
  int    mTitleHeight;
  int    mLayoutType;
